refactor linking services
