+++ apps/mobile/src/hooks/api/providers/strava/useStrava.ts
@@ -3,6 +3,7 @@ import {Linking} from 'react-native';
 import {InAppBrowser} from 'react-native-inappbrowser-reborn';
 import Config from 'react-native-config';
 import api, {getErrors} from '@api';
+import {linkOauth} from '../utils';
 
 const API_URL = Config.API_URL;
 const AUTH_ENDPOINT = 'providers/strava/auth';
@@ -15,27 +16,7 @@ export function useStrava() {
   const link = async () => {
     setLinking(true);
 
-    try {
-      const response = await api.get<any>('/' + AUTH_ENDPOINT);
-
-      if ((await InAppBrowser.isAvailable()) && response.oauth_url) {
-        const decodedUrl = decodeURIComponent(response.oauth_url)
-          .replace(/\n/g, ' ')
-          .replace(/\s+/g, '');
-
-        const browserResponse = await InAppBrowser.openAuth(decodedUrl, '/', {
-          ephemeralWebSession: false,
-        });
-
-        if (browserResponse.type === 'success' && browserResponse.url) {
-          if (!browserResponse.url.includes('auth-success')) {
-            throw Error('Something went wrong.');
-          }
-        }
-      } else Linking.openURL(AUTH_URL);
-    } catch (e) {
-      console.log(e);
-    }
+    await linkOauth(API_URL);
 
     setLinking(false);
   };
